### 2021-11-12 #3906 @whedon
[PRE REVIEW]: PySS3: A new interpretable and simple machine learning model for text classification
**Submitting author:** @sergioburdisso (<a href="http://orcid.org/0000-0002-7694-6834">Sergio Gastón Burdisso</a>)
**Repository:** <a href="https://github.com/sergioburdisso/pyss3" target ="_blank">https://github.com/sergioburdisso/pyss3</a>
**Version:** v0.6.4
