### 2021-11-15 1146c54
Update DOIs of paper references
github.com/openjournals/joss-reviews/issues/3906#issuecomment-967300984
