+++ docs/paper/paper.bib
@@ -23,7 +23,8 @@
   number={1},
   pages={e0168344},
   year={2017},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0168344}
 }
 
 @article{escalante2017early,
@@ -33,7 +34,8 @@
   volume={89},
   pages={99--111},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.eswa.2017.07.040}
 }
 
 @inproceedings{losada2016test,
@@ -42,7 +44,8 @@
   booktitle={International Conference of the Cross-Language Evaluation Forum for European Languages},
   pages={28--39},
   year={2016},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-44564-9_3}
 }
 
 @inproceedings{losada2017erisk,
@@ -51,7 +54,8 @@
   booktitle={International Conference of the Cross-Language Evaluation Forum for European Languages},
   pages={346--360},
   year={2017},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-65813-1_30}
 }
 
 @article{iskandar2017terrorism,
@@ -73,7 +77,7 @@
     issn = "0957-4174",
     author = {Burdisso, Sergio G. and Errecalde, Marcelo and Montes-y-G{\'o}mez, Manuel},
     keywords = "Early text classification, Early depression detection, Incremental classification, SS3, Interpretability, Explainability",
-    doi = {https://doi.org/10.1016/j.eswa.2019.05.023}
+    doi = {10.1016/j.eswa.2019.05.023}
 }
 
 @article{burdisso2019-tss3,
@@ -85,7 +89,7 @@
     issn = "0167-8655",
     author = {Burdisso, Sergio G. and Errecalde, Marcelo and Montes-y-G{\'o}mez, Manuel},
     keywords = "Early text classification, Dynamic word n-grams, Incremental classification, SS3, Explainability, Trie, Digital tree",
-    doi = {https://doi.org/10.1016/j.patrec.2020.07.001}
+    doi = {10.1016/j.patrec.2020.07.001}
 }
 
 @inproceedings{burdisso2019clef,
