### 2021-11-15 #3906 @sergioburdisso
Hi guys! thank you for consider my submission :)

> Hi @sergioburdisso, we will use this issue to find and assign editor and reviewers. While we are doing that, could you resolve those issues with the DOIs in your references?
>
Hi! yes, I've just updated the DOIs and pushed the changes to the repository. :rocket: thank you for the suggestion.

> Note: @hbaniecki has agreed to review, but won't be available for another week or so.

Ok. No problem, so we will wait until @hbaniecki is available then :+1:

Again, thank you guys for taking the time to process and reviewing this submission! :muscle:
