### 2021-11-19 #3934 @whedon
[REVIEW]: PySS3: A new interpretable and simple machine learning model for text classification
**Submitting author:** @sergioburdisso (<a href="http://orcid.org/0000-0002-7694-6834">Sergio Gastón Burdisso</a>)
**Repository:** https://github.com/sergioburdisso/pyss3
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.4
