### 2021-11-23 sergioburdisso/pyss3#17 @hbaniecki
[joss] update the changelog
Hi, I am reviewing your JOSS submission https://github.com/openjournals/joss-reviews/issues/3934 and aim to mention in this repository any potential issues.

I noticed that the `CHANGELOG.md` file (therefore also https://pyss3.readthedocs.io/en/latest/about/changelog.html) is not updated for the versions `0.6.1-0.6.4` that were released to PyPI https://pypi.org/project/pyss3.
