### 2021-11-24 sergioburdisso/pyss3#17 @sergioburdisso
Thank @hbaniecki for opening this issue to inform this, I've just updated the changelong.

And thanks for taking the time to review our submission! I really appreciate it!! :heart_eyes:
