### 2021-11-24 sergioburdisso/pyss3#18 @hbaniecki
[joss] update entry site of the documentation
https://github.com/openjournals/joss-reviews/issues/3934

Hi, I enjoy working with http://tworld.io/ss3, which is highlighted in README. It is also good to see a statement of need with the two references there. Could we add the same to the welcome page of the documentation at https://pyss3.readthedocs.io/en/latest?
