### 2021-11-24 sergioburdisso/pyss3#19 @hbaniecki
[joss] feature request: accessible utility to import a dataset
[openjournals/joss-reviews#3934](https://github.com/openjournals/joss-reviews/issues/3934)

This package has good documentation. Going through the examples I came up with a feature request, which would greatly benefit introducing newcomers and prototyping code.

I like the first example in README to be straightforward and copy-paste ready, which is not the case here (looking at missing code `...`).

How about implementing some `import_dataset(url)` / `download(url)` functionality in `utils` or `Dataset` that would, for example, download the dataset `.zip` file and unpack it ([sample code](https://gist.github.com/hantoine/c4fc70b32c2d163f604a8dc2a050d5f6)) so that one can load the data into exemplary code:
```python
from pyss3 import SS3

Dataset.import_dataset("https://github.com/sergioburdisso/pyss3/blob/master/examples/datasets/movie_review.zip")
x_train, y_train = Dataset.load_from_files("movie_review/train")
x_test, y_test = Dataset.load_from_files("movie_review/test")

clf = SS3()
clf.fit(x_train, y_train)
y_pred = clf.predict(x_test)
```

Implementation details and naming may vary, but it would be nice to easily run code from README.
