### 2021-11-24 95a2a40
Update statement of need in README and Docs (#18)
Resolves: #18
