### 2021-11-24 sergioburdisso/pyss3#18 @sergioburdisso
Hi @hbaniecki! I'm glad to hear you enjoy the live demo! :blush:

Thanks for this suggestion, I've just updated the ``README.md`` to add to more references and also added the same state of need to the welcome page of the documentations as you suggested. Thank you for helping me to improve this project!  :octocat:
