### 2021-11-24 sergioburdisso/pyss3#19 @sergioburdisso
Hi again @hbaniecki!

Wow, this is an amazing idea :clap::clap::clap:

What do you think if adding a method called ``load_from_url()`` to the ``Dataset`` class, which would do the same thing as the current [``Dataset.load_from_files()``](https://pyss3.readthedocs.io/en/latest/api/index.html#pyss3.util.Dataset.load_from_files) but instead of loading the dataset from disk, it will do it from an URL, as you suggested.

Perhaps ``load_from_url()`` should take two arguments, ``load_from_url(url, folder=None)``, first the ``url`` from which to download the zipped dataset and secondly, an **optional** argument called something like ``folder`` that let the user to specify a particular folder to use from inside the zipped dataset. The example would end up being something like:

```Python
from pyss3 import SS3

x_train, y_train = Dataset.load_from_url("https://url/to/movie_review.zip", "train")
x_test, y_test = Dataset.load_from_url("https://url/to/movie_review.zip", "test")

clf = SS3()
clf.fit(x_train, y_train)
y_pred = clf.predict(x_test)
```
What do you think?
(Again, thank you for this suggestion, I think it is an awesome idea :muscle::sunglasses::+1:)
