### 2021-11-25 sergioburdisso/pyss3#19 @hbaniecki
Thanks! If the data won't be saved on disk and only loaded into `x_train, y_train` etc. then `load_from_url(URL, folder=None)` makes perfect sense.
