### 2021-12-04 6a43a35
Allow datasets to be loaded from a given url (#19)
Now datasets can be directly loaded via a given url, not only from disk.
To achieve this, two methods have been added to ``Dataset`` class:

 - ``Dataset.load_from_url(...)``
 - ``Dataset.load_from_url_multilabel(...)``

These methods download and extract the zip file (given by the  url)
into the system's temporary folder and then calls
``Dataset.load_from_files()`` to load it
(or ``Dataset.load_from_files_multilabel()``, respectively).

Note: If the same url is used consecutively, the already downloaded
files will be used as a cache (to avoid
[…]
