### 2021-12-04 4ff35a5
Update README.md according to #19
