### 2021-12-04 sergioburdisso/pyss3#19 @sergioburdisso
Hi @hbaniecki! sorry for the delay, I just had to wait for the weekend to get down on this. I've added the suggested methods and also updated the `README.md`. Just check it out and let me know if everthing is OK :muscle: :nerd_face: :+1:
