### 2021-12-04 ef657e1
Add tests for the methods added to Dataset (#19)
