### 2021-12-05 #3934 @hbaniecki
Overall, the `PySS3` package is mature and of high quality. It comes with extensive documentation and easy-to-follow tutorials. Over two years of development, it steadily gathered a recognizable user-base (55k downloads, 200+ stars). The core functionalities are a valuable and needed contribution to interpretable machine learning, especially in the context of model monitoring and interactive analysis.

**Comment on substantial scholarly effort:**
As highlighted both in the software paper and documentation, two articles of the authors preceded this work:
> "A text classification framework for simple and effective early depression detection over social media streams" (Expert Systems with Applications, 2019), https://doi.org/10.1016/j.eswa.2019.05.023,
"τ-SS3: A text classifier with dynamic n-grams for early risk detection over text streams" (Pattern Recognition Letters, 2020), https://doi.org/10.1016/j.patrec.2020.07.001.

These focus mainly on introducing, evaluating and applying novel algorithms. At the same time, the software (paper) implements a convenient API for the machine learning models and contributes interactive (server, visual) tools to work with them.

**Comment on contribution and authorship**:
From what I can tell looking at the GitHub repository, the submitting author made major contributions to the software, and is the only such in a strict sense of contributing code and documentation. Nevertheless, the full list of paper authors seems appropriate and complete, as the two other authors also contribute to this project, e.g. looking at the two preceding papers.

The authors addressed my issues on **Functionality** and **Documentation**. I will now proceed to review the paper contents.
