### 2022-01-15 sergioburdisso/pyss3#24 @kmichael08
[JOSS] comments on the paper
My comments on the software paper wrt to the JOSS submission:

- In the abstract, you declare two useful tools and then say: "For instance, one of these tools provides (...)". Since there are only two tools, I'd also suggest describing the other one as well. Additionally, instead of "for instance", I'd go with one sentence per each tool. This way, I can only guess what the other functionality is.
- The last sentence in the abstract is quite long, and I would consider breaking it into shorter pieces.
- My understanding is that your input is the implementation of the SS3 algorithm. Therefore, I'd be happy to see a bit more details about the algorithm to make the paper self-contained. Also, the title gives the impression of you introducing a new model/algorithm. I believe the input of this work is an implementation? This distinction is not very clear to me. Also, does it mean the SS3 algorithm was proposed without any implementation? That sounds a bit confusing.
- Is the explanation tool model-specific, or it works with any classification method? What is your exact input here (the explanation algorithm or GUI?). If it's a model-agnostic explanation, perhaps it could be implemented in a different package?
- github -> GitHub?
- since explanations of the models are not the primary contribution of this work (?), you could consider adding a reference to some work in this area.
- Line 40: "On the other hand" doesn't contrast with anything, and definitely no "On the one hand". Maybe you could consider rephrasing this
- Footnote 2 has an issue with spacing  – no space between "ArXiv" and a bracket
- Footnote 2 reference points me to the local file rather than the arXiv preprint

I ticked all the boxes for this part anyways. An exciting paper overall. I like the examples particularly. However, it's not clear what the exact contributions are: model implementation and the explanation GUI (or algorithm)?

In reference to https://github.com/openjournals/joss-reviews/issues/3934
