### 2022-01-15 #3934 @kmichael08
I started with the section related to the software paper. My comments are in the issue linked above.
Just a quick check. This paper is a shorter version of an arXiv preprint from 2019, and this preprint is referenced in this work. @bmcfee is this all good with the JOSS guidelines?
