### 2022-01-15 #3934 @bmcfee
@kmichael08 yes, that should be fine.  The JOSS preprint policy is listed here: https://joss.readthedocs.io/en/latest/submitting.html#preprint-policy

> Authors are welcome to submit their papers to a preprint server (arXiv, bioRxiv, SocArXiv, PsyArXiv etc.) at any point before, during, or after the submission and review process.
>
> Submission to a preprint server is not considered a previous publication.
