### 2022-01-31 #3934 @sergioburdisso
Hi @bmcfee, @hbaniecki @kmichael08 ! First of all, thanks for taking the time to review our paper! I'm having a pretty rough "new year" so far, and couldn't manage to take care of the JOSS issues, but I'll try to do it this weekend, would it be OK if that's the case?
