### 2022-04-04 61c8419
Update JOSS paper (issue #21)
