+++ docs/paper/paper.md
@@ -1,12 +1,12 @@
 ---
-title: 'PySS3: A new interpretable and simple machine learning model for text classification'
+title: 'PySS3: A Python package implementing SS3, a simple and interpretable machine learning model for text classification'
 tags:
   - Python
   - Machine learning
   - Natural Language Processing
   - Text Classification
+  - Interpretability
   - Explainable Artificial Intelligence
-  - XAI
 authors:
   - name: Sergio G. Burdisso^[corresponding author]
     # orcid: 0000-000...
@@ -34,7 +34,7 @@ bibliography: paper.bib
 # Summary
 
 
-In this paper, we briefly introduce `PySS3`[^pyss3] and share it with the community. `PySS3` is an open-source Python package that implements a novel machine learning model for text classification, called SS3. `PySS3` comes with two useful tools that allow working with SS3 in an effortless, interactive, and visual way. For instance, one of these tools provides post hoc explanations using visualizations that directly highlight relevant portions of the raw input document, allowing researchers to better understand the models being deployed. Therefore, `PySS3` could be especially useful for those working with sensitive text classification problems by which people's lives could be affected since it allows researchers and practitioners to deploy interpretable (i.e. self-explainable) and more reliable models for text classification. 
+In this paper, we briefly introduce `PySS3`[^pyss3] and share it with the community. `PySS3` is an open-source Python package that implements the SS3 machine learning model for text classification. `PySS3` comes with useful tools that allow working with SS3 in an interactive, and visual way. For instance, one of these tools provides post hoc explanations using visualizations that directly highlight relevant portions of the raw input document, allowing researchers to better understand the models being deployed. Therefore, `PySS3` could be especially useful for those working with sensitive classification problems by which people's lives could be affected since it allows researchers and practitioners to deploy interpretable (i.e. self-explainable) and more reliable models for text classification. 
 
 [^pyss3]: [https://github.com/sergioburdisso/pyss3](https://github.com/sergioburdisso/pyss3)
 
@@ -57,12 +57,12 @@ Therefore, we decided to develop an open-source Python package to provide the fi
 
 # Implementation
 
-`PySS3` was coded to be compatible with Python 2 and Python 3 as well as with different operating systems, such as Linux, macOS, and Microsoft Windows. To ensure this compatibility holds whenever the source code is updated, we have configured and linked the github repository with the Travis CI service. This service automatically runs the test scripts using different operating systems and versions of Python whenever new code is pushed to the repository.
+`PySS3` was coded to be compatible with Python 2 and Python 3 as well as with different operating systems, such as Linux, macOS, and Microsoft Windows. To ensure this compatibility holds whenever the source code is updated, we have configured and linked the Github repository with the Travis CI service. This service automatically runs the test scripts using different operating systems and versions of Python whenever new code is pushed to the repository.
 
 The package is composed of one main module and three submodules.[^more_info] 
 The main module is called `pyss3` and contains the classifier's implementation _per se_ in a class called `SS3`.
-This class not only implements the "plain-vanilla" version of the classifier [@burdisso2019] but also different variations, such as the one introduced later by the same authors [@burdisso2019-tss3], which allows SS3 to recognize important word n-grams on the fly.
-As the reader will notice in the example shown in the next section, this class exposes a clear and simple API that is similar to that of _Scikit-learn_ models . For instance, it provides standard methods like `fit()` and `predict()` for  training and classification, respectively.[^api_doc]
[…]
