### 2022-04-04 sergioburdisso/pyss3#21 @sergioburdisso
Hi @hbaniecki!

It's been forever, I'm so sorry! I've moved to Switzerland and started working as a postdoctoral researcher here, and it was a huge change in my life.

I've addressed most of the points you highlighted (commit 61c8419), btw THANKS for your valuable advices. Below I'll address each one of your points:

**Comments**
1. The title has been updated following your guidance, now it is "PySS3: A Python package implementing SS3, a simple and interpretable machine learning model for text classification", the word novel has been removed from the paper, as you suggested.
2. This point will be addressed as part of the "statement of need" following the other points suggested by you there.
3. I don't think the API is stable enough yet for a 1.0 version, but I will release a new version with the new changes (including loading dataset from url) and reference to that version in the paper, do you think it is OK? of course if you don't agree we can talk about it, no problem! :)

**Summary**
- Both points were fixed.

**Statement of need**
Yes, I totally agree, in the initial version I didn't include it due to space limitation (in fact the paper exceeded the 1000 words limitation). This point and the ones you pointed out in the following item I'll addressed them both in this section. The idea is to talk about the interpretability and explainability a little bit, cite the papers you suggested, and then add the "gap phrase", like "However, little attention has been paid" etc. and focus on the need of interpretable models (not just explainable, but interpretable, i.e. self-explainable). What do you think?

**State of the field**
These references and this discussion will be added above.

**Implementation**
- All three points have been fixed.

**Illustrative examples**
1. I've added a brief description of the dataset at the beginning.
2. Updated the example using the `Dataset.load_from_url()` function.

**Conclusions**
I've changed the conclusion removing "novel" and adding an extra sentence.

I'm still working on the changes regarding the "Statement of need", I'll let you know as soon as I finish with it. Again, thank you so much for your review work, and apologize for the delay....
