### 2022-04-22 #3934 @bmcfee
Thanks @sergioburdisso for keeping this moving forward - how are things now regarding reviewer comments?

https://github.com/sergioburdisso/pyss3/issues/21
https://github.com/sergioburdisso/pyss3/issues/24
